feat(conditional): Make `#if` unstable as well.
